python/requirements.txt
fastapi
uvicorn
g4f
openai
python-dotenv